moving test-event to common folder, avoiding findBadKeys.js error
